+++ src/db/types.ts
@@ -1,0 +1,10 @@
+export type Question = {
+  id: number;
+  question: string;
+};
+
+export type Answer = {
+  id: number;
+  questionID: number;
+  answer: string;
+};
